refactoring: 공통 컴포넌트 디렉토리 위치 변경
commonModal, CommonTextInput, CommonButtons 들을 CommmonComponent 폴더로 이동
