+++ frontend/guest-app/src/components/CommonComponent/CommonButtons/CancelButton.js
