+++ frontend/guest-app/src/components/CommonComponent/CommonTextInput/CommonTextInput.js
