+++ frontend/guest-app/src/components/CommonComponent/CommonTextInput/useCommonTextInput.js
