+++ frontend/guest-app/src/components/Question/QuestionCard/LikeButton/LikeButton.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Button from "@material-ui/core/Button";
 import ThumbUpIcon from "@material-ui/icons/ThumbUp";
-import useCommonModal from "../../../CommonModal/useCommonModal.js";
+import useCommonModal from "../../../CommonComponent/CommonModal/useCommonModal.js";
 
 import UndoLikeConfirmModal from "./UndoLikeModal.js";
