+++ frontend/guest-app/src/components/Question/QuestionCard/LikeButton/UndoLikeModal.js
@@ -2,9 +2,9 @@ import React from "react";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import {Box} from "@material-ui/core";
-import CommonModal from "../../../CommonModal/CommonModal.js";
-import ConfirmButton from "../../../CommonButtons/ConfirmButton.js";
-import CancelButton from "../../../CommonButtons/CancelButton.js";
+import CommonModal from "../../../CommonComponent/CommonModal/CommonModal.js";
+import ConfirmButton from "../../../CommonComponent/CommonButtons/ConfirmButton.js";
+import CancelButton from "../../../CommonComponent/CommonButtons/CancelButton.js";
 
 function UndoLikeConfirmModal({isOpened, onCancelClick, onConfirmClick}) {
 	return (
